+++ crates/ruma-appservice-api/src/ping/send_ping.rs
@@ -9,14 +9,16 @@ pub mod unstable {
 
     use ruma_common::{
         OwnedTransactionId,
-        api::{auth_scheme::AccessToken, request, response},
+        api::{request, response},
         metadata,
     };
+
+    use crate::HomeserverToken;
 
     metadata! {
         method: POST,
         rate_limited: false,
-        authentication: AccessToken,
+        authentication: HomeserverToken,
         path: "/_matrix/app/unstable/fi.mau.msc2659/ping",
     }