+++ crates/ruma-appservice-api/src/thirdparty/get_user_for_user_id.rs
@@ -9,15 +9,17 @@ pub mod v1 {
 
     use ruma_common::{
         OwnedUserId,
-        api::{auth_scheme::AccessToken, request, response},
+        api::{request, response},
         metadata,
         thirdparty::User,
     };
 
+    use crate::HomeserverToken;
+
     metadata! {
         method: GET,
         rate_limited: false,
-        authentication: AccessToken,
+        authentication: HomeserverToken,
         path: "/_matrix/app/v1/thirdparty/user",
     }