+++ crates/ruma-client-api/src/admin/is_user_locked.rs
@@ -14,7 +14,7 @@ pub mod v1 {
 
     use ruma_common::{
         OwnedUserId,
-        api::{auth_scheme::AccessToken, request, response},
+        api::{OAuthClientScope, auth_scheme::AccessToken, request, response},
         metadata,
     };
 
@@ -22,6 +22,12 @@ pub mod v1 {
         method: GET,
         rate_limited: false,
         authentication: AccessToken,
+        required_scopes: [
+            #[cfg(not(feature = "unstable-msc4484"))]
+            OAuthClientScope::ApiFullAccess,
+            #[cfg(feature = "unstable-msc4484")]
+            OAuthClientScope::ServerAdministration,
+        ],
         history: {
             unstable("uk.timedout.msc4323") => "/_matrix/client/unstable/uk.timedout.msc4323/admin/lock/{user_id}",
             1.18 => "/_matrix/client/v1/admin/lock/{user_id}",