+++ crates/ruma-identity-service-api/src/association/email/validate_email.rs
@@ -9,14 +9,16 @@ pub mod v2 {
 
     use ruma_common::{
         OwnedClientSecret, OwnedSessionId,
-        api::{auth_scheme::AccessToken, request, response},
+        api::{request, response},
         metadata,
     };
+
+    use crate::IdentityServiceToken;
 
     metadata! {
         method: POST,
         rate_limited: false,
-        authentication: AccessToken,
+        authentication: IdentityServiceToken,
         history: {
             1.0 => "/_matrix/identity/v2/validate/email/submitToken",
         }