+++ crates/ruma-identity-service-api/src/lib.rs
@@ -7,6 +7,8 @@
 
 #![warn(missing_docs)]
 
+use ruma_common::api::auth_scheme::{AuthScheme, ExtractTokenError, add_access_token_as_authorization_header, extract_bearer_or_query_token};
+
 pub mod association;
 pub mod authentication;
 pub mod discovery;
@@ -16,3 +18,33 @@ pub mod lookup;
 pub mod tos;
 
 ruma_common::priv_owned_str!();
+
+/// Authentication is required by including an identity server access token in the
+/// `Authentication` http header, or an `access_token` query parameter.
+/// 
+/// Using the query parameter is deprecated since Matrix 1.11.
+#[derive(Debug, Clone, Copy, Default)]
+pub struct IdentityServiceToken;
+
+impl AuthScheme for IdentityServiceToken {
+    type Input<'a> = &'a str;
+    type AddAuthenticationError = http::header::InvalidHeaderValue;
+    /// The identity service token.
+    type Output = String;
+    type ExtractAuthenticationError = ExtractTokenError;
+
+    fn add_authentication<T: AsRef<[u8]>>(
+        request: &mut http::Request<T>,
+        access_token: Self::Input<'_>,
+    ) -> Result<(), Self::AddAuthenticationError>
+    {
+        add_access_token_as_authorization_header(request.headers_mut(), access_token)
+    }
+
+    fn extract_authentication<T: AsRef<[u8]>>(
+        request: &http::Request<T>,
+    ) -> Result<Self::Output, Self::ExtractAuthenticationError>
+    {
+        extract_bearer_or_query_token(request)?.ok_or(ExtractTokenError::MissingAccessToken)
+    }
+}