+++ crates/ruma-identity-service-api/src/tos/accept_terms_of_service.rs
@@ -8,14 +8,16 @@ pub mod v2 {
     //! [spec]: https://spec.matrix.org/v1.18/identity-service-api/#post_matrixidentityv2terms
 
     use ruma_common::{
-        api::{auth_scheme::AccessToken, request, response},
+        api::{request, response},
         metadata,
     };
+
+    use crate::IdentityServiceToken;
 
     metadata! {
         method: POST,
         rate_limited: false,
-        authentication: AccessToken,
+        authentication: IdentityServiceToken,
         history: {
             1.0 => "/_matrix/identity/v2/terms",
         }